75: edelta.personlist.example -> edelta.examples
Task-Url: http://github.com/LorenzoBettini/edelta/issues/75
