+++ edelta.parent/edelta.examples/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>edelta.personlist.example</name>
+	<name>edelta.examples</name>
 	<comment></comment>
 	<projects>
 	</projects>
