+++ edelta.parent/edelta.examples/.settings/org.eclipse.core.resources.prefs
