+++ edelta.parent/edelta.examples/.settings/org.eclipse.jdt.core.prefs
