+++ edelta.parent/edelta.examples/.settings/org.eclipse.m2e.core.prefs
