+++ edelta.parent/edelta.examples/META-INF/MANIFEST.MF
@@ -1,10 +1,10 @@
 Manifest-Version: 1.0
-Automatic-Module-Name: edelta.personlist.example
+Automatic-Module-Name: edelta.examples
 Bundle-ManifestVersion: 2
-Bundle-Name: edelta.personlist.example
+Bundle-Name: edelta.examples
 Bundle-Vendor: Lorenzo Bettini
 Bundle-Version: 0.9.0.qualifier
-Bundle-SymbolicName: edelta.personlist.example; singleton:=true
+Bundle-SymbolicName: edelta.examples;singleton:=true
 Bundle-ActivationPolicy: lazy
 Require-Bundle: edelta.lib,
  edelta.refactorings.lib,
@@ -15,5 +15,3 @@ Require-Bundle: edelta.lib,
 Import-Package: org.apache.log4j,
  org.apache.commons.logging
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Export-Package: com.example,
- edelta.personlist.example.tests
