+++ edelta.parent/edelta.examples/build.properties
