+++ edelta.parent/edelta.examples/edelta-gen/edelta/personlist/example/ExampleSyntax1.java
@@ -1,4 +1,4 @@
-package com.example;
+package edelta.personlist.example;
 
 import edelta.lib.AbstractEdelta;
 import edelta.refactorings.lib.EdeltaRefactorings;
