+++ edelta.parent/edelta.examples/model/PersonList.ecore
