+++ edelta.parent/edelta.examples/modified-syntax1/PersonList.ecore
