+++ edelta.parent/edelta.examples/modified-syntax2/PersonList.ecore
