+++ edelta.parent/edelta.examples/modified-syntax2/README
