+++ edelta.parent/edelta.examples/pom.xml
@@ -7,7 +7,7 @@
     <artifactId>edelta.parent</artifactId>
     <version>0.9.0-SNAPSHOT</version>
   </parent>
-  <artifactId>edelta.personlist.example</artifactId>
+  <artifactId>edelta.examples</artifactId>
   <packaging>eclipse-test-plugin</packaging>
 
   <build>
